app/requirements.txt
pydantic-settings
fastapi
line-bot-sdk==3.16.0
openai==1.64.0
uvicorn
